fix(plugin.ts): Updated code and restructured plugin
Updated code and restructured plugin
