+++ .babelrc
@@ -1,3 +1,4 @@
 {
-  "presets": [ "react", "env", "stage-0"]
+  "presets": [ "react", "env", "stage-0"],
+  "plugins": ["babel-plugin-react-native-web"]
 }
