+++ .storybook/webpack.config.js
@@ -1,9 +1,34 @@
 // load the default config generator.
-const genDefaultConfig = require('@storybook/react/dist/server/config/defaults/webpack.config.js');
+//const genDefaultConfig = require('@storybook/react/dist/server/config/defaults/webpack.config.js');
 const path =require('path');
-module.exports = (baseConfig, env) => {
-  const config = genDefaultConfig(baseConfig, env);
+const appDirectory = path.resolve(__dirname, '../');
 
+module.exports = (config, env) => {
+	
+ config.module.rules.push({
+	
+	test: /\.(js)$/,
+	 // Add every directory that needs to be compiled by Babel during the build.
+	 include: [
+		path.resolve(appDirectory, 'node_modules/react-native-vecto-icons/lib/create-icon-set'),
+		 path.resolve(appDirectory, 'node_modules/react-native-vecto-icons'),
+		path.resolve(appDirectory, 'node_modules/react-native-vecto-icons'),
+	//	path.resolve(appDirectory, 'src'),
+  ],
+	  use: {
+		loader: 'babel-loader',
+		options: {
+			cacheDirectory: true,
+			// Babel configuration (or use .babelrc)
+			// This aliases 'react-native' to 'react-native-web' and includes only
+			// the modules needed by the app.
+			plugins: ['react-native-web'],
+			//How to add loader in module of webpack The 'react-native' preset is recommended to match React Native's packager
+			presets: ['react-native']
+		}
+	}
+  });
+	
 	config.module.rules.push(
 		{
 			test: /\.ttf$/,
@@ -14,8 +39,10 @@ module.exports = (baseConfig, env) => {
 		test: /\.(ts|tsx)$/,
 		loader: require.resolve('awesome-typescript-loader')
 	});
-
-	config.resolve.extensions.push('.ts', '.tsx');
+	console.log("Config",config.module.rules);
+	
+	config.resolve.alias={  'react-native': 'react-native-web'}
+	config.resolve.extensions.push('.web.js', '.ts', '.tsx','.js','.ttf','.jsx');
 
 	return config;
 };
