+++ src/FontAwesomeIcons/index.web.tsx
