--- src/FontAwesomeIcons/package.json
@@ -1,153 +1,0 @@
-{
-  "name": "bluerain-plugin-vector-icons",
-  "version": "0.0.0",
-  "description": "",
-  "keywords": [],
-  "main": "dist/bluerain-plugin-vector-icons.umd.js",
-  "module": "dist/bluerain-plugin-vector-icons.es5.js",
-  "typings": "dist/types/bluerain-plugin-vector-icons.d.ts",
-  "files": [
-    "dist"
-  ],
-  "author": "BlueEast Team <code@blueeast.com>",
-  "repository": {
-    "type": "git",
-    "url": ""
-  },
-  "license": "MIT",
-  "engines": {
-    "node": ">=6.0.0"
-  },
-  "scripts": {
-    "lint": "tslint -t codeFrame 'src/**/*.ts' 'test/**/*.ts'",
-    "prebuild": "rimraf dist",
-    "build": "tsc && babel -D compiled --out-dir dist  && rimraf compiled",
-    "start": "tsc -w & rollup -c rollup.config.ts -w",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:prod": "npm run lint && npm run test -- --coverage --no-cache",
-    "deploy-docs": "ts-node tools/gh-pages-publish",
-    "report-coverage": "cat ./coverage/lcov.info | coveralls && codecov",
-    "commit": "git-cz",
-    "semantic-release": "semantic-release pre && npm publish && semantic-release post",
-    "semantic-release-prepare": "ts-node tools/semantic-release-prepare",
-    "precommit": "lint-staged",
-    "storybook": "start-storybook -p 6006",
-    "storybook-native": "storybook start -p 7007",
-    "prestorybook-native": "rnstl",
-    "build-storybook": "build-storybook",
-    "android": "react-native-scripts android",
-    "ios": "react-native-scripts ios"
-  },
-  "bundlesize": [
-    {
-      "path": "./dist/bluerain-plugin-vector-icons.es5.js",
-      "maxSize": "3 kB"
-    }
-  ],
-  "lint-staged": {
-    "{src,test}/**/*.ts": [
-      "prettier --write --use-tabs --single-quote",
-      "git add"
-    ]
-  },
-  "config": {
-    "commitizen": {
-      "path": "node_modules/cz-conventional-changelog"
-    },
-    "validate-commit-msg": {
-      "types": "conventional-commit-types",
-      "helpMessage": "Use \"npm run commit\" instead, we use conventional-changelog format :) (https://github.com/commitizen/cz-cli)"
-    },
-    "react-native-storybook-loader": {
-      "searchDir": [
-        "./src"
-      ],
-      "pattern": "**/*.stories.tsx",
-      "outputFile": "./storybook/storyLoader.tsx"
-    }
-  },
-  "jest": {
-    "transform": {
-      ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
-    },
-    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js"
-    ],
-    "coveragePathIgnorePatterns": [
-      "/node_modules/",
-      "/test/"
-    ],
-    "coverageThreshold": {
-      "global": {
-        "branches": 90,
-        "functions": 95,
-        "lines": 95,
-        "statements": 95
-      }
-    },
-    "collectCoverage": true,
-    "mapCoverage": true
-  },
-  "devDependencies": {
-    "@blueeast/bluerain-cli": "^1.2.1",
-    "@blueeast/bluerain-platform-reactxp": "^1.1.8",
-    "@blueeast/tslint-config-blueeast": "^0.4.0",
-    "@storybook/addon-actions": "^3.3.12",
-    "@storybook/addon-graphql": "^3.2.0",
-    "@storybook/addon-info": "^3.2.9",
-    "@storybook/addon-knobs": "^3.3.13",
-    "@storybook/addon-links": "^3.3.12",
-    "@storybook/addon-storyshots": "^3.2.13",
-    "@storybook/addons": "^3.2.13",
-    "@storybook/react": "^3.3.12",
-    "@types/expo__vector-icons": "6.2.1",
-    "@types/jest": "^21.1.0",
-    "@types/react-native-vector-icons": "^4.4.3",
-    "@types/storybook__react": "^3.0.6",
-    "awesome-typescript-loader": "^3.4.1",
-    "bundlesize": "^0.15.2",
-    "codecov": "^2.3.1",
-    "colors": "^1.1.2",
-    "commitizen": "^2.9.6",
-    "coveralls": "^3.0.0",
-    "cross-env": "^5.0.1",
-    "cz-conventional-changelog": "^2.0.0",
-    "husky": "^0.14.0",
-    "jest": "^21.0.0",
-    "lint-staged": "^6.0.0",
-    "lodash.camelcase": "^4.3.0",
-    "prettier": "^1.4.4",
-    "prompt": "^1.0.0",
-    "prop-types": "^15.6.0",
-    "react-dom": "^16.2.0",
-    "react-native-scripts": "1.8.1",
-    "react-native-storybook-loader": "^1.7.0",
-    "react-native-typescript-transformer": "^1.2.3",
-    "replace-in-file": "^3.0.0-beta.2",
-    "rimraf": "^2.6.1",
-    "rollup": "^0.52.0",
-    "rollup-plugin-commonjs": "^8.0.2",
-    "rollup-plugin-node-resolve": "^3.0.0",
-    "rollup-plugin-sourcemaps": "^0.4.2",
-    "semantic-release": "^8.0.0",
-    "storybook-addon-intl": "^2.1.2",
-    "storybook-addon-rtl": "^0.1.1",
-    "ts-jest": "^22.0.0",
-    "tslint": "^5.4.3",
-    "typedoc": "^0.9.0",
-    "typescript": "^2.3.4",
-    "validate-commit-msg": "^2.12.2"
-  },
-  "dependencies": {
-    "@blueeast/bluerain-os": "3.5.3",
-    "@expo/vector-icons": "^6.3.1",
-    "expo": "^25.0.0",
-    "react": "16.2.0",
-    "react-native": "0.44.0",
-    "react-native-vector-icons": "4.0.0"
-  }
-}
