+++ src/index.tsx
@@ -1,6 +1,6 @@
 
-import {VectorIonicons} from './VectorIonIcons/';
-import {VectorMaterialIcons} from  './MaterialIcons/';
-import {VectorFontAwesome} from  './FontAwesomeIcons/';
+import {VectorIonicons} from './VectorIonIcons/index.web';
+import {VectorMaterialIcons} from  './MaterialIcons/index.web';
+import {VectorFontAwesome} from  './FontAwesomeIcons/index.web';
 export {VectorIonicons,VectorMaterialIcons,VectorFontAwesome};
